+++ src/Applet/Start/appletStartCommand.ts
@@ -1,41 +1,50 @@
-import { CommandLineOptions } from "command-line-args";
-import ICommand from '../../Command/ICommand';
 import { createEmulator } from '../../Emulator/emulatorFactory';
-import { ENTRY_FILE_PATH_OPTION } from '../Upload/appletUploadCommand';
 import {
 	getAppletEntryFileAbsolutePath,
 	getAppletDirectoryAbsolutePath as getProjectDirAbsolutePath,
 	getAppletEntryFileRelativePath,
+	APPLET_PATH_OPTION,
+	ENTRY_FILE_PATH_OPTION,
 } from '../Upload/appletUploadCommandHelper';
 import { getAppletDirAbsolutePath } from './appletStartCommandHelper';
 import { loadEmulatorOrCreateNewAndReturnUid } from '../../Emulator/emulatorFacade';
+import { CommandLineOptions, createCommandDefinition } from '../../Command/commandDefinition';
+import { NO_DEFAULT_ORGANIZATION_OPTION, ORGANIZATION_UID_OPTION } from '../../Organization/organizationFacade';
 
-export const appletStart: ICommand = {
+export const OPTION_LIST = [
+	NO_DEFAULT_ORGANIZATION_OPTION,
+	ORGANIZATION_UID_OPTION,
+	{ name: 'port', type: Number, description: `Port where will the applet run`, defaultValue: 8090 },
+	APPLET_PATH_OPTION,
+	{ name: 'project-dir', type: String, description: 'Directory of the applet project' },
+	ENTRY_FILE_PATH_OPTION,
+] as const;
+
+export const appletStart = createCommandDefinition({
 	name: 'start',
 	description: 'Start applet locally',
-	optionList: [
-		{ name: 'port', type: Number, description: `Port where will the applet run`, defaultValue: 8090 },
-		{ name: 'applet-dir', type: String, description: 'Directory of the applet to start' },
-		{ name: 'project-dir', type: String, description: 'Directory of the applet project' },
-		ENTRY_FILE_PATH_OPTION,
-	],
+	optionList: OPTION_LIST,
 	commands: [],
-	async run(options: CommandLineOptions) {
+	async run(options: CommandLineOptions<typeof OPTION_LIST>) {
 		const currentDirectory = process.cwd();
 
+		if (!options.port) {
+			throw new Error('Argument --port is required');
+		}
 		const emulatorServerPort = options.port;
 		const entryFileAbsolutePath = await getAppletEntryFileAbsolutePath(currentDirectory, options);
 		const projectDirAbsolutePath = await getProjectDirAbsolutePath(currentDirectory, options);
 		const appletDirAbsolutePath = await getAppletDirAbsolutePath(currentDirectory, options);
 		const entryFileRelativePath = await getAppletEntryFileRelativePath(entryFileAbsolutePath, appletDirAbsolutePath);
-		const emulatorUid = await loadEmulatorOrCreateNewAndReturnUid();
+		const emulatorUid = await loadEmulatorOrCreateNewAndReturnUid(options);
 
-		await createEmulator({
+		const createEmulatorParams = {
 			emulatorUid,
 			appletPath: appletDirAbsolutePath,
 			projectPath: projectDirAbsolutePath,
 			entryFileRelativePath,
 			emulatorServerPort,
-		});
+		};
+		await createEmulator(createEmulatorParams, options);
 	},
-};
+});