+++ src/Applet/Test/Upload/appletTestUploadCommand.ts
@@ -1,51 +1,50 @@
 import chalk from 'chalk';
 import * as prompts from 'prompts';
-import { CommandLineOptions } from 'command-line-args';
-import ICommand from '../../../Command/ICommand';
 import { createOrganizationRestApi, } from '../../../helper';
-import { getOrganization, ORGANIZATION_UID_OPTION } from '../../../Organization/organizationFacade';
+import { getOrganization, getOrganizationUidOrDefaultOrSelect, NO_DEFAULT_ORGANIZATION_OPTION, ORGANIZATION_UID_OPTION } from '../../../Organization/organizationFacade';
 import { loadTestFilesContents, validateTestFiles } from './appletTestUploadFacade';
 import {
+	getAppletUid,
 	getAppletVersion,
-	tryGetAppletUid,
 } from '../../appletFacade';
-import {
-	getOrganizationUidAndUpdateConfig,
-} from '../../Upload/appletUploadCommandHelper';
 import { createProgressBar } from '../../../CommandLine/progressBarFactory';
 import { loadPackage } from '../../../FileSystem/packageConfig';
 import IAppletTestSuite from '@signageos/sdk/dist/RestApi/Applet/Version/IAppletTestSuite';
+import { CommandLineOptions, createCommandDefinition } from '../../../Command/commandDefinition';
 
-export const appletTestUpload: ICommand = {
+const OPTION_LIST = [
+	NO_DEFAULT_ORGANIZATION_OPTION,
+	ORGANIZATION_UID_OPTION,
+	{
+		name: 'yes',
+		type: Boolean,
+		description: `Allow to upload new applet test or override existing test without confirmation step`,
+	},
+	{
+		name: 'verbose',
+		type: Boolean,
+		description: `outputs all files to upload`,
+	},
+] as const;
+
+export const appletTestUpload = createCommandDefinition({
 	name: 'upload',
 	description: 'Uploads applet test',
-	optionList: [
-		ORGANIZATION_UID_OPTION,
-		{
-			name: 'yes',
-			type: Boolean,
-			description: `Allow to upload new applet test or override existing test without confirmation step`,
-		},
-		{ // will output all files to upload for multifile applet
-			name: 'verbose',
-			type: Boolean,
-			description: `outputs all files to upload`,
-		},
-	],
+	optionList: OPTION_LIST,
 	commands: [],
-	async run(options: CommandLineOptions) {
+	async run(options: CommandLineOptions<typeof OPTION_LIST>) {
 		const isVerbose = !!options.verbose;
 		const skipConfirmation = !!options.yes;
 
 		const currentDirectory = process.cwd();
-		const organizationUid = await getOrganizationUidAndUpdateConfig(options);
+		const organizationUid = await getOrganizationUidOrDefaultOrSelect(options);
 		const organization = await getOrganization(organizationUid);
 		const restApi = createOrganizationRestApi(organization);
 
 		const version = await getAppletVersion(currentDirectory);
-		let appletUid = await tryGetAppletUid(currentDirectory);
+		const appletUid = await getAppletUid(restApi);
 		if (!appletUid) {
-			throw new Error(`applet uid is not present in package file.`);
+			throw new Error('Not selected Applet or sos.appletUid is not present in package.json');
 		}
 
 		const applet = await restApi.applet.get(appletUid);
@@ -119,7 +118,7 @@ export const appletTestUpload: ICommand = {
 			progressBar.end();
 		}
 	},
-};
+});
 
 function displaySuccessMessage(
 	appletName: string,