+++ src/Applet/appletCommand.ts
@@ -1,10 +1,10 @@
-import ICommand from "../Command/ICommand";
 import { appletGenerate } from './Generate/appletGenerateCommand';
 import { appletUpload } from './Upload/appletUploadCommand';
 import { appletStart } from './Start/appletStartCommand';
 import { appletTest } from "./Test/appletTestCommand";
+import { createCommandDefinition } from '../Command/commandDefinition';
 
-export const applet: ICommand = {
+export const applet = createCommandDefinition({
 	name: 'applet',
 	description: 'Applet management',
 	optionList: [],
@@ -17,4 +17,4 @@ export const applet: ICommand = {
 	async run() {
 		throw new Error('Unknown command');
 	},
-};
+});