+++ src/Device/PowerAction/powerActionCommand.ts
@@ -1,22 +1,24 @@
-import ICommand from "../../Command/ICommand";
-import { getOrganization, getOrganizationUid, ORGANIZATION_UID_OPTION } from "../../Organization/organizationFacade";
-import { DEVICE_UID_OPTION, getDeviceUid, typeMap } from "../deviceFacade";
-import { CommandLineOptions } from "command-line-args";
+import { getOrganization, getOrganizationUidOrDefaultOrSelect, NO_DEFAULT_ORGANIZATION_OPTION, ORGANIZATION_UID_OPTION } from "../../Organization/organizationFacade";
+import { DEVICE_UID_OPTION, getDeviceUid, POWER_ACTION_TYPE_OPTION, typeMap } from "../deviceFacade";
 import { createOrganizationRestApi } from "../../helper";
 import chalk from "chalk";
 import { getActionType } from "../deviceFacade";
+import { CommandLineOptions, createCommandDefinition } from "../../Command/commandDefinition";
 
-export const powerAction: ICommand = {
+const OPTION_LIST = [
+	NO_DEFAULT_ORGANIZATION_OPTION,
+	ORGANIZATION_UID_OPTION,
+	DEVICE_UID_OPTION,
+	POWER_ACTION_TYPE_OPTION,
+] as const;
+
+export const powerAction = createCommandDefinition({
 	name: 'power-action',
 	description: 'Perform power action on device',
-	optionList: [
-		ORGANIZATION_UID_OPTION,
-		DEVICE_UID_OPTION,
-		{ name: 'type', type: String, description: `Type of device power action`},
-	],
+	optionList: OPTION_LIST,
 	commands: [],
-	async run(options: CommandLineOptions) {
-		const organizationUid = await getOrganizationUid(options);
+	async run(options: CommandLineOptions<typeof OPTION_LIST>) {
+		const organizationUid = await getOrganizationUidOrDefaultOrSelect(options);
 		const organization = await getOrganization(organizationUid);
 		const restApi = createOrganizationRestApi(organization);
 		const deviceUid = await getDeviceUid(restApi, options);
@@ -27,4 +29,4 @@ export const powerAction: ICommand = {
 			console.log(chalk.green(`Action ${typeMap.get(actionType)!.name} was successful on device ${deviceUid}`));
 		});
 	},
-};
+});