+++ src/Device/deviceFacade.ts
@@ -1,13 +1,13 @@
 import * as Debug from 'debug';
 import * as prompts from 'prompts';
-import { CommandLineOptions } from "command-line-args";
 import { deserializeJSON, postResource } from '../helper';
 import { IOrganization } from '../Organization/organizationFacade';
-import { getGlobalApiUrl } from '../Command/commandProcessor';
 import { DevicePowerAction } from '@signageos/sdk/dist/RestApi/Device/PowerAction/IPowerAction';
 import RestApi from "@signageos/sdk/dist/RestApi/RestApi";
 import { IApplet } from "../Applet/appletFacade";
-import { getMachineIp } from "./Connect/connectHelper";
+import { CommandLineOptions } from '../Command/commandDefinition';
+import { getGlobalApiUrl } from '../Command/globalArgs';
+import { getMachineIp } from '../Helper/localMachineHelper';
 
 const debug = Debug('@signageos/cli:Device:facade');
 
@@ -35,11 +35,12 @@ export const typeMap = new Map<string, ActionData>(
 	],
 );
 
-export const DEVICE_UID_OPTION = { name: 'device-uid', type: String, description: 'Device UID' };
+export const DEVICE_UID_OPTION = { name: 'device-uid', type: String, description: 'Device UID' } as const;
+export const POWER_ACTION_TYPE_OPTION = { name: 'type', type: String, description: `Type of device power action` } as const;
 
 export async function getDeviceUid(
 	restApi: RestApi,
-	options: CommandLineOptions,
+	options: CommandLineOptions<[typeof DEVICE_UID_OPTION]>,
 ) {
 	let deviceUid: string | undefined = options['device-uid'];
 	if (!deviceUid) {
@@ -62,7 +63,7 @@ export async function getDeviceUid(
 	return deviceUid;
 }
 
-export async function getActionType(options: CommandLineOptions)  {
+export async function getActionType(options: CommandLineOptions<[typeof POWER_ACTION_TYPE_OPTION]>)  {
 	let action: string | undefined = options.type;
 
 	if (!action) {