+++ src/Emulator/emulatorFactory.ts
@@ -9,8 +9,9 @@ import * as fsExtra from 'fs-extra';
 import * as glob from 'globby';
 import chalk from 'chalk';
 import { IEmulator } from './IEmulator';
-import { loadConfig } from '../RunControl/runControlHelper';
 import { createDomain } from './createDomain';
+import { getOrganizationUidOrDefaultOrSelect, NO_DEFAULT_ORGANIZATION_OPTION, ORGANIZATION_UID_OPTION } from '../Organization/organizationFacade';
+import { CommandLineOptions } from '../Command/commandDefinition';
 
 export interface ICreateEmulatorParams {
 	projectPath: string;
@@ -33,7 +34,10 @@ type IEnvVars = {
 	checksum: string;
 };
 
-export async function createEmulator(params: ICreateEmulatorParams): Promise<IEmulator> {
+export async function createEmulator(
+	params: ICreateEmulatorParams,
+	options: CommandLineOptions<[typeof ORGANIZATION_UID_OPTION, typeof NO_DEFAULT_ORGANIZATION_OPTION]>,
+): Promise<IEmulator> {
 	const { projectPath, emulatorServerPort, appletPath, entryFileRelativePath } = params;
 	const entryFileAbsolutePath = path.join(appletPath, entryFileRelativePath);
 
@@ -43,8 +47,7 @@ export async function createEmulator(params: ICreateEmulatorParams): Promise<IEm
 
 	const packageConfig = JSON.parse(fsExtra.readFileSync(path.join(projectPath, 'package.json')).toString());
 
-	const sosGlobalConfig = await loadConfig();
-	const organizationUid = sosGlobalConfig.defaultOrganizationUid;
+	const organizationUid = await getOrganizationUidOrDefaultOrSelect(options);
 
 	if (!organizationUid) {
 		throw new Error(`No default organization selected. Use ${chalk.green('sos organization set-default')} first.`);