+++ src/Organization/Get/organizationGetCommand.ts
@@ -1,21 +1,23 @@
 import chalk from 'chalk';
 import * as Debug from 'debug';
-import { CommandLineOptions } from "command-line-args";
-import { getOrganizationUid, getOrganization, ORGANIZATION_UID_OPTION } from '../organizationFacade';
-import ICommand from '../../Command/ICommand';
+import { CommandLineOptions, createCommandDefinition } from '../../Command/commandDefinition';
+import { getOrganization, getOrganizationUidOrDefaultOrSelect, NO_DEFAULT_ORGANIZATION_OPTION, ORGANIZATION_UID_OPTION } from '../organizationFacade';
 const debug = Debug('@signageos/cli:Organization:get');
 
-export const organizationGet: ICommand = {
+const OPTION_LIST = [
+	NO_DEFAULT_ORGANIZATION_OPTION,
+	ORGANIZATION_UID_OPTION,
+] as const;
+
+export const organizationGet = createCommandDefinition({
 	name: 'get',
 	description: 'Get one organization by UID',
-	optionList: [
-		ORGANIZATION_UID_OPTION,
-	],
+	optionList: OPTION_LIST,
 	commands: [],
-	async run(options: CommandLineOptions) {
+	async run(options: CommandLineOptions<typeof OPTION_LIST>) {
 		debug('Organization get');
-		const organizationUid = await getOrganizationUid(options);
+		const organizationUid = await getOrganizationUidOrDefaultOrSelect(options);
 		const organization = await getOrganization(organizationUid);
 		console.log(chalk.yellow(JSON.stringify(organization, undefined, 2)));
 	},
-};
+});