+++ src/Timing/List/timingListCommand.ts
@@ -1,23 +1,25 @@
 import chalk from 'chalk';
 import * as Debug from 'debug';
-import { CommandLineOptions } from "command-line-args";
-import { getOrganizationUid, getOrganization, ORGANIZATION_UID_OPTION } from '../../Organization/organizationFacade';
+import { getOrganization, getOrganizationUidOrDefaultOrSelect, NO_DEFAULT_ORGANIZATION_OPTION, ORGANIZATION_UID_OPTION } from '../../Organization/organizationFacade';
 import { getDeviceUid, DEVICE_UID_OPTION } from '../../Device/deviceFacade';
 import { createOrganizationRestApi } from '../../helper';
-import ICommand from '../../Command/ICommand';
+import { CommandLineOptions, createCommandDefinition } from '../../Command/commandDefinition';
 const debug = Debug('@signageos/cli:Timing:list');
 
-export const timingList: ICommand = {
+const OPTION_LIST = [
+	NO_DEFAULT_ORGANIZATION_OPTION,
+	ORGANIZATION_UID_OPTION,
+	DEVICE_UID_OPTION,
+] as const;
+
+export const timingList = createCommandDefinition({
 	name: 'list',
 	description: 'List timings assigned to device',
-	optionList: [
-		ORGANIZATION_UID_OPTION,
-		DEVICE_UID_OPTION,
-	],
+	optionList: OPTION_LIST,
 	commands: [],
-	async run(options: CommandLineOptions) {
+	async run(options: CommandLineOptions<typeof OPTION_LIST>) {
 		debug('Timing create');
-		const organizationUid = await getOrganizationUid(options);
+		const organizationUid = await getOrganizationUidOrDefaultOrSelect(options);
 		const organization = await getOrganization(organizationUid);
 		const restApi = createOrganizationRestApi(organization);
 		const deviceUid = await getDeviceUid(restApi, options);
@@ -26,4 +28,4 @@ export const timingList: ICommand = {
 		});
 		console.log(chalk.yellow(JSON.stringify(timings, undefined, 2)));
 	},
-};
+});