+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- The appletUid does not have to be hardcoded inside package.json and is auto-detected from current organization based on name
+- Support for profiles inside the ~/.sosrc file using ini `[profile xxx]` sections and SOS_PROFILE env. var. or `--profile` argument 
+
+### Fixed
+- When default organization is not set it asks for saving it to the current ~/.sosrc file
+
+### Changed
+- The option `--no-update-package-config` is reversed into option `--update-package-config` and by default the package.json is not updated. See README.
+- The `defaultOrganizationUid` is now always used as default for all commands instead of selecting one. Use argument `--no-default-organization` or remove line `defaultOrganizationUid` from `~/.sosrc` to prevent this.
+
 ## [0.10.3] - 2022-01-18
 ### Fixed
 - Compatibility with peer dependency for front-display version 9.13.0+ (because of changed API)